Address comments, refactor splitGrepResultsByFile into grepSearch implementation
